Pass modified date to tar file header
If `gulp-tar` is run multiple times on the same content the resulting tar will have different contents. This is because the default behavior of the `archiver` library is to use the current date if none have been provided for the entry header.

With this change the modified date from the file itself will be provided allowing for the same `tar` file to be consistently created from the same contents.